Update services filenames & move all auth to handler
